backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-migrate==4.0.5
pymongo==4.5.0
redis==4.6.0
chromadb==0.4.15
langchain==0.0.335
langchain-groq==0.1.5
groq==0.4.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
celery==5.3.4
pandas==2.0.3
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
bcrypt==4.1.2
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
